# specs/002-ai-native-textbook-platform/contracts/api.yaml
openapi: 3.1.0
info:
  title: Physical AI Textbook API
  description: Backend API for the AI-Native Textbook Platform
  version: 1.0.0
  contact:
    name: Panaversity
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://physical-ai-api.onrender.com
    description: Production (Render)
paths:
...
  /api/chat:
    post:
      summary: Send a chat message to the Professor
      operationId: chat
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Chat service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chat/history:
    get:
      summary: Get chat history for authenticated user
      operationId: getChatHistory
      tags: [Chat]
      security:
        - sessionToken: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Chat history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatHistoryItem'
                  total:
                    type: integer
        '401':
          description: Not authenticated
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: cookie
      name: session_token
  schemas:
...
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          maxLength: 2000
          example: What is a ROS 2 Node?
        page_context:
          type: string
          description: Current page path for context-aware responses
          example: /docs/module-01-ros2/nodes
        user_profile:
          type: object
          description: Optional user profile for personalization (anonymous users)
          properties:
            software_skills:
              type: array
              items:
                type: string
            hardware_inventory:
              type: array
              items:
                type: string
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
          description: The AI-generated response
        sources:
          type: array
          items:
            type: object
            properties:
              chunk_id:
                type: string
              source_page:
                type: string
              snippet:
                type: string
        personalized:
          type: boolean
          description: Whether the response was personalized to user profile
    ChatHistoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message_text:
          type: string
        response_text:
          type: string
        page_context:
          type: string
        created_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string